Clear pending invalidations in lookahead when removed from lookahead
This change unsets the flags tracking lookahead meausurement and
placement invalidations when a node is removed from the LookaheadScope.
When using movableContent to move layout subtrees from one parent
in a LookaheadScope to another parent not in a LookaheadScope, we
need to make sure we don't carry over the invalidation flags for
lookahead, lookahead is no longer valid/needed in that case.

Fixes: 392014142
Test: New test added
Change-Id: I38993fd488a284e3f91f47ef2f0387ddc1048a56