Add parameter for Runtime Mediation Integration Testing
It adds support for end to end testing
of Runtime Mediation use cases. This adds callbacks
to the SdkRuntime Mediatee SDK inside the test
app, which triggers callbacks to the test
process. Also, adds parameterized approach to the
UiPresentationTests file for testing all mediation cases
in just a single trigger to run the tests.

Bug: 395070351
Test: UiPresentationTests
Change-Id: I3cde27fb946d8e289c3b5699771dca4b42a85c7c