Do not allow null values in backstackMap
If the savedState for a destination is `null`, we don't have to save
that state into the backstackMap. We can keep an emptyString
instead and when we restore the state convert it back to null.

RelNote: "Fixed an issue where attempting to saveState with
non-inclusive pop would result in a null savedState that could cause a
crash on restoration."
Test: added test
Bug: 392112864
Bug: 395091644

Change-Id: I9f3e8d272db222041c7e611932e4c2b7b4fa50c1