Fix exception when moving AndroidView in nested MC
A previous workaround introduced a delay in detaching an AndroidView
from its parent, which led to a situation where we attempted to reparent
an already attached view when the AndroidView moved inside of a nested
MovableContent block.

The workaround here stemmed from b/240139002, which I'm not able to
reproduce after making this change.

Fixes: b/384865168
Test: asNestedMovableContentChild, asNestedMovableContentChildWithReuse
Change-Id: Ib40cfeb23f5fd66e3a5ab882d055ea25f45bcbbc