Introduce MetadataTestHelper
We are removing public Metadata constructor.
Considering some tests may rely on the constructor to fill in the auto populated
fields in metadata, here in the test library we provide a way to
populate those values, to support those use cases.

In our fake implementation, also enable developers to set the package
name. For the same purpose.

Bug: 394783116
Test: ./gradlew :health:connect:connecting-testing:test
Relnote: Introduce MetadataTestHelper
Change-Id: I1f7f1bd207cadaaaf8a75fbb74bbe986986af9fc