Fix: Zoom, scroll and swipe were not working properly with ViewPager
The scroll events were being consumed by the ViewPager library thus
creating a conflict with the GestureTracker in the library. We need to
let the host app handle the horizontal fling gesture and not the scroll
gesture. Thus in this CL, the GestureTracker explicitly disallows the
ancestors of PdfView to handle touch if until the content is at the
edges in the viewport

Bug: 394512887
Bug: 394807115
Bug: 394814877
Bug: 395534901

Test: Manual as V2 fragment is not migrated to V1. Automation test is
created under V1

Change-Id: I7e57a12b5dc6de43a234598b636b2b05fe0d2272