[Webkit] Add maxPrerenders to SpeculativeLoadingConfig API
Add handling for maxPrerenders for SpeculativeLoadingConfig API. There
was an error in WebView so we have to update the Feature value to move
away from the broken WebView versions in the newer AndroidX version.

Remove the MAX_* numbers from SpecualativeLoading as WebView now can
limit them.

Bug: 382048812
Test: Updated PrefetchTest
Relnote: "Add maxPrerenders config to SpeculativeLoadingConfig API"
Change-Id: I67245b2f422f413b92829f7f6ea234bfd9390f74