Add handle variants to TraceProcessor APIs to allow custom lifecycle without block callbacks
Bug: 388801063
Test: TraceProcessorTest

Relnote: Added TraceProcessor and Session API with closable Handles, for easier usage with custom lifecycles. This is also a step towards enabling easier Coroutine and Java API usage. The extension functions to TraceProcessor.runServer {} are now marked experimental, as they're likely to move, and be made non-experimental in the future.

Change-Id: I358b4256110d2b0c4b35bb9f0be4ae2f407f0791