Use PrivacySandboxCallback instead of aidl for in-app mediatee.
This is being done to avoid having both aidl and shim in the test app.

We are introducing a runtime aware layer that enables communication between the RE_SDK and the inapp mediatee.
This layer (inappmediatee adapter) implements a PrivacySandboxCallback which is registered with the RE_SDK by the app.

Test: tested on test app

Bug: 388775437
Change-Id: Ied0ff6a6fbc2b39a5670a8048cc13ad81dfdea8d