Handle missing annotation default values in non-JVM projects
For non-JVM projects using KSP1, annotation values that are not explicitly defined will be null instead of being the annotation default value. This CL handles these by filtering null values when wrapping KSValueArgument and swapping them with the default value. The behaviour of an argument having its default value if not explicitly declared in use-site is part of KSP but to workaround this issue this CL also implements it in XProcessing. Note that null is not a valid annotation value.

Testing is only possible outside androidx since xprocessing-testing uses K2JVMCompiler while Room's KMP integration test app uses KSP2, both scenarios where the issue is not reproducible.

Bug: 329508813
Bug: 396607230
Test: Manual in KMP project with Kotlin 2.1.10 and KSP with useKsp2 = false
Change-Id: I883b890f6e844ef3efd351ed7c5cf3004969e44c