Ensure state is saved until pop animation complete
SaveableStateNavLocalProvider needs to ensure that even though keys are
remove from the back stack, if their content are still in composition,
we do not destroy that state until after the composition is complete.

This means that we need ProvideToEntry to participate in the clean up
and to ensure we can handle duplicate keys properly.

This does not consider the cases where developers could potentially call
ProvideToEntry on duplicate keys out of order (i.e. call on a entry
whose key is not the last copy on the back stack). We will need further
API considerations to handle that.

Test: all tests pass
Bug: 392184079
Change-Id: If41f3e33ef45381bfde13fce511639a6c84b63cc