Enable loading SQLite JNI from configured path.
Configuring the system property 'androidx.sqlite.driver.bundled.path' will make it so that BundledSQLiteDriver's native library is loaded from the configured location instead of extracting it from the JAR and loading it from a temporary file. This allows for bundled / sandboxed apps to configure the loading from the bundle directory.

Bug: 381282544
Test: BundledSQLiteDriverTest
Change-Id: Ic8c01d74bf3e61edbabbfbdc9b24ca79f4dcd3a9