Graphics low latency: use blendMode=SRC when copying previous frame's buffer to next frame
The behavior of using HardwareRenderer.isOpaque is tested by existing code in
PreservedBufferContentsVerifier, and if it finds that a device doesn't preserve
the contents of the previous draw, then it is copying over the contents in full
from one frame to the next. On most devices, it seems that if content isn't
preserved in the buffer, then the buffer is cleared. However, it seems that some
devices put unexpected data in the buffer - for Daylight DC-1, that content seems
to be every pixel with 0xffc0c0c0, at least with Theme.AppCompat applied to the
activity (the behavior seems to be theme-dependent). To avoid this unexpected
data appearing, use blendMode=SRC instead of the default blendMode=SRC_OVER, so
that even areas without any content being copied over will be cleared instead of
drawn on top of unexpected data.

This fixes the behavior on LowLatencyCanvasView, as well as the Ink Authoring module.

Test: Verified with Pixel Tablet, Daylight DC-1 (where this issue manifests), and Samsung S7+.

Change-Id: I51bbd5eaa43efb405fc59d47c08b23101b792e46