Fixes selection menu placement on Samsung devices
Previously, we were hosting the ActionMode in the PageMosaicView that
contains the selection, and we were specifying the coordinates to
place the menu in PageMosaicView coordinate space (i.e. relative to the
page, without accounting for any scale / scroll applied by the parent
ZoomView). This worked fine on Pixel devices, but it was empirically
buggy on Samsung devices. Adjusting the menu coordinates to account
for scale applied via ZoomView would fix the menu placement on Samsung
devices but break it on Pixel devices. It follows that Pixel and
Samsung implementations of floating action mode have different
treatment of onGetContentRect when an ActionMode is hosted in a child
View with scale applied via a parent ViewGroup.

The fix is to host the ActionMode directly in the scaled parent
(ZoomView), and to supply menu coordinates relative to ZoomView. This
side steps the different OEM treatments when trying to host and place
the menu within the child PageMosaicView.

Demos of the fixed behavior can be found in the linked bug.

This fixes several other bugs related to selection menu placement:
1) Fixes the assumption that the entire selection exists within the
rectangle defined by the top left corner of the start handle and the
bottom right corner of the end handle. This is not true in many cases,
particularly those that concern multi-line selections. This was fixed
by examining the total selection bounds instead of the selection handle
locations to determine menu visibility / placement in both
SelectionActionMode and ZoomScrollValueObserver.
2) Fixes the assumption that ZoomView and children fill the entire
width of the display. This was fixed by using ZoomView's
getVisibleAreaInContentCoords in lieu of Resources#getDisplayMetrics
to compute which part of the selection is visible. The former reports
actual layout values from the View hierarchy while the latter reports
the physical dimensions of the screen. There are many cases where a
View hierachy does not fill the screen.
3) Fixes a minor, outright typo in the V2 (PdfView) implementation

Test: manual, ./gradlew :pdf:pdf-viewer:(test|cAT)
Bug: 395520781
Change-Id: Ie6cd11fd0fab2d97b477aa412f2b300c65f99e10