Add launcher to create non-LifecycleOwner sandboxed activities
Adds LocalUnmanagedSdkActivityLauncher for activities that do not extend
the LifecycleOwner interface. It extends the existing
SdkActivityLauncher interface. To create a launcher, an extension method
createUnmanagedSdkActivityLauncher is added for Activity.

As opposed to LocalManagedSdkActivityLauncher, the new launcher must be
manually disposed by the user.

Relnote: "Added a launcher for non-LifecycleOwner activities. Renamed
LocalSdkActivityLauncher to LocalManagedSdkActivityLauncher."
Bug: 382012571
Test: CreateSdkActivityLauncherTest

Change-Id: Ib14089762f10b1ae42989b6ef66191bdae4beb06