Changing the XR Stubs from PUBLISHED to SNAPSHOT_ONLY library.
This is so we can remove the API requirements and the @RestrictTo annotations with it, such that the produced AAR can be consumed internally. This library is only used to satisfy SceneCore's compilation requirement (i.e. it's not being released) and is not the source of truth for the API surface (that would be Android XR).

Test: Annotation-only change.
Fixes: 399207163
Change-Id: I253bf8c3bfa9e99f802fa4a41b90cdd41926c299