Rework layout node invalidation order
This CL refined the invalidation order.

Previously we required lookahead invalidations to happen before all
non-lookahead invalidations. While this works for most cases, there
are use cases where a lookahead invalidation of a node is handled
before its lookaheadRoot is measured/placed. This is because
lookaheadRoot's ancestor far up the tree (out of LookaheadScope)
would be invalidated after lookahead invalidations based on this
order.

This CL refined the invalidation order, such that nodes that are
out of LookaheadScope have the highest priority in handling
invalidations. This creates a guarantee that all lookaheadRoots'
ancestors are invalidated and updated before the invalidations
reach the LookaheadScopes.

Test: new test added
Fixes: 383821644
Fixes: 393149313
Change-Id: I0497c8d840862b5c7634dc85a91d456722b53710