Add new API CustomTabsIntent.Builder#setCloseButtonEnabled and relevant extra flag.
This CL introduceds a new API to CustomTabsIntent.Builder to enable or
disable the close button in custom tab, it's enabled by default, i.e.
shown in the toolbar of the tab. For captive portal using custom tab
case, the close button doesn't make sense, the tab should be closed
automatically once the network is connected and validated. This API
will provide a way for us to disable the close button in the tab.

Bug: 399545305
Test: ./gradlew updateApi
Test: ./gradlew browser:browser:test
Relnote: Add new API to enable/disable the close button for custom tab.
Change-Id: I35acd34edac715e380d64ccb1c7e71385cda53b2