Fix Transition not animating when provided with a new TransitionState
The remember statements within animateValue would seem like it's capturing the changes in targetState and segment, however, it's also capturing the initial Transition<S> object.

Meaning that when providing a new MutableTransitionState to rememberTransition, the new targetState and segment would be ignored, causing no animation to be run on the new Transition object. Effectively leaking the old Transition<S> object into animateValue.

Bug: 399478091
Test: TransitionTest.kt
Change-Id: Ibea099d0e1243b47bdaa48191206d90e4317b2c0