Updating PdfViewerFragmentV2 to match the already exposed API surface of V1.
This CL updates the getters and setters of PdfView to restrict the
visibility. Getters and setters had to be explicitly defined as
`RestrictTo` on property was not working. Because of this, lateinit
modifier had to removed as kotlin does not allow it. Hence the protected
property had to be made nullable. This property is used for testing
purposes.

Introduced a private property _pdfView for more control.

Bug: 398787913
Test: ./gradlew :pdf:pdf-viewer-fragment:checkApi
Change-Id: I1208f0e6093162cd08b35cfd8fb4e7a47e080435