Replace SparseParcelableArraySerializer with SparseArraySerializer
The current implementation requires users to annotate the SparseArray as well as the element type with a @Serializable annotation when the element Parcelable has no associated serializer. This element serializer is only used to satisfy the compiler and is not used in any other way. Note the repetition of "Parcelable" in both the container serializer and the element serializer:

```
@Serializable(with = SparseParcelableArraySerializer::class)
val value: SparseArray<@Serializable(with = MyParcelableSerializer::class) MyParcelable>
```

This is bad as users may use our SparseParcelableArraySerializer with SparseArrays of other element types without any compile-time errors or warnings, e.g.

```
@Serializable(with = SparseParcelableArraySerializer::class)
val value: SparseArray<String>
```

So we're replacing SparseParcelableArraySerializer with a more general SparseArraySerializer. Now the container serializer does the container part and the element serializer does the element part. Users can use it in a more idiomatic way:

```
@Serializable(with = SparseArraySerializer::class)
val value: SparseArray<@Serializable(with = MyParcelableSerializer::class) MyParcelable>
```

This also enables users to serialize SparseArray of other element types, e.g.

```
@Serializable(with = SparseArraySerializer::class)
val value: SparseArray<String>
```

Test: SavedStateCodecAndroidTest
Relnote: Replace SparseParcelableArraySerializer with SparseArraySerializer
Change-Id: I91de872ff549b0262418cdb9c1e4d682c69f84ee