Add Android PendingIntent as a supported AppFunctionDatatype with metadata.
This cl: Adds pending intent to the supported types in app functions.
Pending intents are mainly supported as a capability in an `OpenIntent` interface but there is not general limitations in writing other schemas that take pending intents in the future.

RelNote: "Add Android PendingIntent as a supported AppFunctionDatatype with metadata"
Bug: 400422328
Test: ./gradlew appfunctions:appfunctions-compiler:test
Change-Id: I540de17f1a0033180b2805f1dc757201a72faf06