Fix material-adaptive for Kotlin 2.1.20
Compose compiler will convert functions in final classes to be restartable and skippable.

DefaultAnimatedPaneOverride#AnimatedPane now has a restartable scope under new rules and AnimatedVisibility was not affecting the caller that provided transitions. The transitions in the caller were calculated based on  PaneMotionData#motion, but it was not using a backing state. It was mostly a coincidence that composition was not skipping any of the AnimatedPane layers before and recomputing transitions correctly.

Fixes: 400512633
Test: Verified that ThreePaneScaffoldMotionScreenshotTest passes with 2.1.20.
Change-Id: Ic3f77b8b203ff290672fdc50c497fa28904a9333