[Webkit] Add Prefetch callback handling
Instead of passing the value type in the boundary interface,
we can just pass an enum with the needed information and let
the client side (AndroidX) create the appropriate exception
 using the information passed.

This way should be more efficient and easier to maintain.
This CL adds the handling for the new boundary interface method.

Test: N/A
Relnote: Updated the Prefetch Feature name.
Change-Id: I606fd63c5425e96d196a72d5469c05a893669ac6