Make projectOrArtifact Gradle Isolated Projects safe.
In groovy, a call to projectOrArtifact triggers a property search in the
project and all of its ancestors, before it finally tries ext["projectOrArtifact"].
This causes a project isolation violation.

Moving to androidx.projectOrArtifact avoids reaching out to parent projects.

Note, it is still broken in playground, that will have to be follow up
work.

Bug: 358448803
Test: ./gradlew projects -Dorg.gradle.unsafe.isolated-projects=true -Dorg.gradle.internal.isolated-projects.configure-on-demand.tasks=true
      fewer warnings in the report
Change-Id: I5321069b22f15de30a2b2c46f1c9f60dd9f00def