Quick fix to verify if this fixes b/401046148
In aosp/3465279 we store focus using a placeholder view, and then
restore focus after onEndApplyChanges. This makes views behave
differently, and it doesn't match the compose behavior where
removing the focused item results in us clearing focus.

This CL adds a clearFocus() to the placeholder view before
it is removed from the hierarchy.

Bug: 401046148
Test: Can't repro this in a test, will write a test if this fixes the issue
Relnote: Added a flag around the logic added in aosp/3465279
Change-Id: I59ea216cb17a889d5c2a3a99ad739fef5d0ea3f0