Fix aliasing issue with (all equal) rounded corners on elements with Modifiers.Background corners
Our M3 components (among other elements), often have rounded
corners which have aliasing issue. This is because we only use
the recommended ViewOutlineProvider for fully circular shapes,
not for all.

This fixes that issue, by applying shape only in
ProtoLayoutInflater on the BackgroundDrawable by clipping its
outline to the correct shape. The diff is noticeable, even in the
screenshot tests.

Move test with corners to full screen one so that clipping is applied
correctly in test.

Fixes: 357061501
Test: Manually + screenshots
Change-Id: I488bd7245dfb5848e930c9e87c625b09a9873891