[402282799] Fix IllegalArgumentException while creating PendingIntent
Fix IllegalArgumentException while creating PendingIntent.

Recommendation AndroidX library creates a PendingIntent inside
getNotificationObject function that doesn't follow latest Android
recommendation of defining the mutability state of it which causes an
IllegalArgumentException thrown by the system.

In this gerrit, I define the mutability to be mutable, as PendingIntents
are assumed to be mutable by default up until Android 11 to ensure
backward compatibility.

Fixes: b/402282799

Bug: 402282799

Test: Built the code using gradle

Change-Id: Id1c6812e485b3aaf0d6c06760f522c35967855fe