[ButtonGroup] API Review Feedback
Update button group to have a animateWidth modifier that developers will need to pass to button group's children for width animation to happen.
This is replacing the approach of automatically passing up interaction sources from children to button group.

Test: Update button group unit and screenshot tests to use this new animateWidth modifier for the children.
Bug: b/391877101
Relnote: Adding animateWidth modifier in ButtonGroupScope that will need to be used with button group's children to correctly animate the children.
Change-Id: Ia3bb63650b255054524581f6fece6c759beff45a