Set the SurfaceView of SandboxedSdkView default mode as Z-Below
Set the default mode of the SandboxedSdkView as Z-Below while
SandboxedSdkView#orderProviderUiAboveClientUi(true) can be used to set
it to z-above.

The toogle from the manual test app also changed to be in z-below mode
in default and change the mode to z-above when enabled.

Relnote: "The default mode of SandboxedSdkView changed to set provider's
surface below the client window."

Bug: 403238319
Bug: 345688233
Test: SandboxedSdkViewTest
Change-Id: I9f18afe4563fb76e327381ab9a07516abed2446b