uiautomator: Ensure UiDevice#getWindowRoots returns roots in window order
UiAutomation#getWindows returns windows in the Z-order but
UiDevice#getWindowRoots did not guarantee the Z-order of returned roots.
This change addresses the inconsistency to ensure that the returned
list of the root is sorted in descending window order, with the
topmost window's root listed first.
This will be useful when multiple UI elements are matched for a search
criteria particularly in multi-instance multi-window mode. User can
easily predict how they are aligned in window stack on screen.

Test: ./gradlew :test:uiautomator:uiautomator:cAT (UiDeviceTest)
Bug: 397565934
Change-Id: I874266f868d541420dde15cce2d240523dbf5276