Clear focus when the focused item is detached
When the focused item is detached, we schedule invalidation
after onEndApplyChanges. However, when a lazylist item is
removed, onDetached() runs after onEndApplyChanges, and the
invalidation waits for the next onEndApplyChanges to take
effect.

This isn't usually a problem since deleting the focused item
changes the UI which would end with another onEndApplyChanges.
However this is an issue if the focused item that is being
removed is part of a beyondBoundsLayout. In this case there
is no UI change that would trigger another onEndApplyChanges.

We fix this issue by clearing focus whenever a node is reset
and don't wait for it to be detached.

Test: Added a new test
Bug: 404290007
Relnote: "We now clear focus when a focusable item is marked
for reuse before it is removed from a lazy list. We would
previously wait until the item was detached."
Change-Id: I43cacea1028429b8e546b0eeeff1b1123e4d0369