SupportSQLiteDatabase wrapper of RoomDatabase
The API RoomDatabase.getSupportWrapper() will return a SupportSQLiteDatabase wrapper from a RoomDatabase even if the database is configured with a driver.

These are APIs are meant to be used as a migration path for codebases that extensively use SupportSQLite APIs with Room (via roomDatabase.openHelper.writableDatabase) but want to migrate to using SQLiteDriver APIs but are unable to completely refactor all their codebase.

Test: SupportSQLiteDatabaseWrapperTest
Relnote: Add APIs to get a SupportSQLiteDatabase wrapper of RoomDatabase with a configured SQLiteDriver.
Change-Id: Icf6ac2215f2b5f36b0a79081d431fd31df9e5901