Decoupled zoom restoration logic from scroll to position
Restoring zoom on config change was coupled with scroll position to restore. While it may work fine for scenarios like view re-creation, scenario where view is retained and it's layout changes, zoom will not be measured correctly.

We partly relied on onConfigChange for adjust zoom, but inside viewpager, onConfigChange may not be called for adjacent view.

This CL cater in the fact that zoom need to be adjusted only on re-creation or on size change of view. For other scenarios like setting initial zoom, getDefaultZoom() would be fine.

Bug: 402031732
Test: Manual
Change-Id: I4874e6910e5e4e22e91601d75e2876e3b4e5bcae