Add a gradle nightly integration test
This creates a github action that will once per day build the collection project using the latest release-nightly build of gradle.

It adds a `gradle-version` input to the existing `build-single-project` action used by the presubmit workflow, which defaults to the gradle wrapper version so that the presubmit workflow will keep using the existing version.

The gradle nightly workflow will optionally apply a patch file from `.github/integration-patches/gradle-nightly.patch` before running the build, to be used when there's a breaking change in gradle which requires an AndroidX code change, but the change can't be committed until AndroidX is actually updated to the latest gradle version.

Bug: 369355667

Test: Sample runs from changes applied to a branch of the github (set to run at push instead of nightly for testing):
  * AndroidX Presubmits workflow still works with the version of gradle from gradlew: https://github.com/androidx/androidx/actions/runs/13980687483
  * Gradle Nightly Integration Test runs with latest gradle nightly: https://github.com/androidx/androidx/actions/runs/13980687488, the build fails due to the removal of an internal API which seems to be used by a jetbrains API (logs here https://github.com/androidx/androidx/actions/runs/13980687488/job/39145065379)
  * Gradle Nightly Integration Test with patch applied fails at a different line: search "benchmark/benchmark-darwin/build.gradle" in this log compared to the one linked above https://github.com/androidx/androidx/actions/runs/13980733878/job/39145207765

Change-Id: I72116448cf0b10678872d7956aecf07e29893325