Fix pop state & predictive back animations
Duplicate backstacks were considered a pop when it should not be. This is an internal implementation detail and it didn't used to be reflected in animations (because duplicate stacks / keys are not animated), so it wasn't discovered until we started implementing zIndex.

Because of this bug, the existing test testNavDisplayPredictiveBackAnimations was passing when it shouldn't be. After the bug was fixed, testNavDisplayPredictiveBackAnimations started failing. And this helped catch a second bug where predictive back was using navigate animations rather than pop animations.

Bug: 391699685
Bug: 393605277
Test: existing tests pass
Change-Id: I85848d2c294a055fe66fddba0ce73922bb876a08