Prevent overflow with beyondViewportPageCount in Pager measurement
This is an alternate fix to 394760799 that allows for
beyondViewportPageCount to still be arbitrarily large.
beyondViewportPageCount is now coerced to be at most pageCount before
doing any layout calculations with it, which should prevent overflow
wherever it's used.

Test: Added test with beyondViewportPageCount = Int.MAX_VALUE and
page > 0 to repro a known overflow case. Also added a pager test config
so all tests run with beyondViewportPageCount = Int.MAX_VALUE.
Fixes: 394760799

Change-Id: Idb2dbbe5e8d24bb032664c02a3b3fffd6a8d3845