FocusRestorer should not pin items
This reverts aosp/3084070 which pined the item that was
marked for focus restoration.

Users should use a key to ensure that the correct item
is restored.

Bug: 330696779
Test: Added new test
Relnote: "FocusRestorer no longer pins the previously
focused item. Users should use a key to ensure that
the previously focused item has the same composition
hash, so that focus is successfully restored."
Change-Id: I4203b9aede7ed93d78a58fd330a807384142dc48