Apply SandboxedSdkView's alpha to it's contentView
View framework doesn't apply a parent view's alpha to its child automatically,
and SurfaceViews have a non-standard treatment of alpha.
We need to explicitly handle this in the UI lib, by changing our SandboxedSdkView.setAlpha()
logic to apply the same alpha value to the contentView.

This'll fix the problem we have of changing the SandboxedSdkView's
alpha not causing any visual change on the UI presented.

Unused button resizeFromSdk in Resize CUJ is also removed.

Bug: b/407048879
Test: Manual CUJ
Change-Id: Ie14b2bce47105301bcaa251ac57ab1cd1fcb85cb