Increase website compatibility in payment demo
Before this patch, if a PaymentRequestActivity user navigated to a 3p
website for testing the payment APIs more broadly, these websites had a
high chance of not working correctly, because of disabled features in
WebView.

This patch updates PaymentRequestActivity with:
1) Enable several features in  based on the WebView shell app example.
2) Make links open in the same WebView.
3) Make the Back button map to navigation within WebView, if possible.
4) Handle orientation changes without loosing JavaScript state.

After this patch, 3p websites are easier to test for payment app
interoperability, because the PaymentRequestActivity is compatible with
more websites, as well as supporting device orientation changes
gracefully.

Test: PaymentRequestActivity
Bug: 408192949
Change-Id: I239492ec2ff0dcbed968bf62697513c3a0ef465c